refactor: StructuredDataRetriever에서 파라미터 검증 로직 분리
StructuredDataRetriever가 파라미터 추출, 날짜 검증, 템플릿 조회,
쿼리 실행, 동적 SQL fallback, 결과 포맷팅 등 6가지 책임을 갖고 있었다.

- QueryParameterValidator 추출: 날짜 검증/보정, 템플릿 존재 확인 로직 분리
- StructuredDataRetriever: 데이터 조회 오케스트레이션에 집중 (140줄→100줄)
- 중복 리터럴 "(결과 없음)" 을 EMPTY_RESULT 상수로 통합
